API/requirements.txt
pandas
ipywidgets
numpy
ipyleaflet
shapely
geojson
geemap
geopy
tifffile
matplotlib
imageio
flask
notebook
pytest
multiprocess
retry
pydantic
fastapi
uvicorn[standard]
python-dateutil
httpx